+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/adap/mgfexport/AdapMgfExportModule.java
@@ -14,7 +14,7 @@
  * 02111-1307, USA.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.adapmgfexport;
+package net.sf.mzmine.modules.peaklistmethods.io.adap.mgfexport;
 
 import java.util.Collection;
 import javax.annotation.Nonnull;
