+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/adap/mgfexport/AdapMgfExportParameters.java
@@ -14,7 +14,7 @@
  * 02111-1307, USA.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.adapmgfexport;
+package net.sf.mzmine.modules.peaklistmethods.io.adap.mgfexport;
 
 import net.sf.mzmine.parameters.Parameter;
 import net.sf.mzmine.parameters.impl.SimpleParameterSet;
