+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/adap/mspexport/AdapMspExportTask.java
@@ -14,7 +14,7 @@
  * 02111-1307, USA.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.mspexport;
+package net.sf.mzmine.modules.peaklistmethods.io.adap.mspexport;
 
 import java.io.File;
 import java.io.FileWriter;
